[NF-8840] Travis build hook for Rails 4 compat
See lib/ci/README.md for more details. This creates a script that
watches for Travis builds to complete. When one completes, if it's a PR
it will count the number of failures, errors and passing tests for each
job that allows failures (Rails 4 jobs). It then compares those against
the equivalent jobs from the last build to finish against the master
branch. Then it takes the results and posts a comment to the GitHub pull
request with the results. If the pull request makes any sub-build worse,
it labels the PR as Rails 4 Rejected. If it improves the situation or
leaves it the same it will label it Rails 4 Accepted. If the master
build and feature branch build do not have the same number of
sub-builds, it labels it as needing to be rebased against master.

Deploying this to GCE or running it locally requires you to set the
GITHUB_TOKEN environment variable with a personal access token that you
must setup. See:
https://help.github.com/articles/creating-an-access-token-for-command-line-use/