Errors were being saved between runs
The error instance variable was being saved on the listener class, this
class isn't re-initialized on each run it sits and listens for jobs.
This means that we ended up keeping errors around when they were out of
date.

There are a few things I think we should do to this listener class.
1. Split it up. It is doing too many things. Not only does it listen for
new jobs it also validates, compares, and comments on each job. Each
thing should probably be it's own class
2. Write tests. This class isn't that well tested and I suspect if we
wrote tests first the class wouldn't have grown to be so large.